feat: move input states back into the input.rs class
