--- dora-explorer/src/library/iobuf/mod.rs
@@ -1,12 +1,0 @@
-use tui_input::Input;
-
-
-pub enum InputBuffer {
-    Active(Input),
-    Inactive,
-}
-
-pub enum OutputBuffer {
-    Normal(String),
-    Error(String),
-}
