+++ dora-explorer/src/library/state.rs
@@ -24,10 +24,6 @@ pub struct ExplorerState{
     pub view_slice: [u16;2],
     available_area: [u16;2], // height, width
 
-    // buffer states
-    input_buffer: InputBuffer,
-    output_buffer: OutputBuffer,
-
     // trap signals
     // when these gets flagged, it will exit.
     pub sig_user_input_exit: bool,
@@ -72,8 +68,6 @@ impl ExplorerState {
             cursor_y: 0,
             view_slice: [0,10], // this will be overridden very quickly
             available_area: [10, 10], // to be reset very soon.
-            input_buffer: InputBuffer::Inactive,
-            output_buffer: OutputBuffer::Normal("".to_string()),
             sig_user_input_exit: false,
             sig_file_selected_exit: false,
         }
@@ -97,8 +91,6 @@ impl ExplorerState {
             cursor_y: 0,
             view_slice: [0,10], // this will be overridden very quickly
             available_area: [10, 10], // to be reset very soon.
-            input_buffer: InputBuffer::Inactive,
-            output_buffer: OutputBuffer::Normal("".to_string()),
             sig_user_input_exit: false,
             sig_file_selected_exit: false,
         }
